refactor: extract admin settings actions
